Added media qmodule + tweaks in pic qmodule
